requirements.txt
Django==3.2
psycopg2>=2.8
djangorestframework==3.13.1
django-cors-headers==3.11.0
